+++ Programacion/eclipse/calculadora/src/calculadora/Suma.java
@@ -4,10 +4,10 @@ public class Suma extends Operaciones
 {
 	public static void sum(String num1,String num2)
 	{
-		int n1,n2,result=0;
+		double n1,n2,result=0;
 		String resultado;
-		n1=Integer.parseInt(num1);
-		n2=Integer.parseInt(num2);
+		n1=Double.parseDouble(num1);
+		n2=Double.parseDouble(num2);
 		result=n1+n2;
 		resultado= String.valueOf(result);
 		setOperacion(resultado);
